+++ CHANGELOG.md
@@ -8,6 +8,11 @@ client
   * relax stage name length restriction from 30 to 240 characters (because domains in san certificates can be up to 256 characters long).
   * set container hostname to md5(app + stage) + dockerhost to make sure the name does not exceed the 64 character limit of the kernel
 
+server
+* added per-stage environment, clients configure it with `stool config environment=FOO:bar` - note the colon to separate key and value;
+  explicit environment arguments passed to start overwrite per-stage environment values
+
+
 ### 5.0.2 (2019-08-29)
 
 client